refactor: split oversized files to pass 400-line CI gate
Split 5 files that exceeded the 400-line limit:
- domain-analysis-cache.ts (663→273) → +cache-types.ts, +snapshot-builder.ts
- shared.ts (592→381) → +domain-analysis-aggregation.ts, +domain-analysis-scoring.ts
- types.ts (500→393) → +types-detail.ts
- analyze-basic.ts (442→290) → +analyze-basic-data.ts
- DomainAnalysis.tsx (410→350) → +domainAnalysisUtils.ts

Public APIs unchanged; all imports updated.
