+++ cloud/apps/api/src/graphql/queries/domain/shared.ts
@@ -1,8 +1,7 @@
 import { db, resolveDefinitionContent } from '@valuerank/db';
 import { formatVnewLabel, formatVnewSignature } from '@valuerank/shared/trial-signature';
 import { DOMAIN_ANALYSIS_VALUE_KEYS, type DomainAnalysisValueKey, extractValuePair } from '../domain-analysis-values.js';
-import { resolveTranscriptDecisionModel } from './decision-model.js';
-import type { TranscriptDecisionModelResult } from './decision-model.js';
+import type { DomainAnalysisVignetteDetail } from './types-detail.js';
 import { runMatchesSignature, selectDefaultVnewSignature } from './planning-utils.js';
 export { selectLatestDefinitionPerLineage, hydrateDefinitionAncestors } from '../../../services/definition-lineage.js';
 export type { LineageDefinitionRow } from '../../../services/definition-lineage.js';
@@ -32,6 +31,13 @@ export type {
 } from './decision-model.js';
 
 export { formatVnewLabel, formatVnewSignature };
+export { aggregateValueCountsFromTranscripts, incrementPairwiseWin, isDomainAnalysisValueKey } from './domain-analysis-aggregation.js';
+export { computeFullBTScores, computeSmoothedLogOddsScore } from './domain-analysis-scoring.js';
+export type {
+  DomainAnalysisConditionDetail,
+  DomainAnalysisConditionTranscript,
+  DomainAnalysisVignetteDetail,
+} from './types-detail.js';
 
 export const MAX_LIMIT = 500;
 export const DEFAULT_LIMIT = 50;
@@ -81,36 +87,6 @@ export type DomainAnalysisMissingDefinition = {
   missingAllModels: boolean;
   missingModelIds: string[];
   missingModelLabels: string[];
-};
-
-export type DomainAnalysisConditionDetail = {
-  scenarioId: string | null;
-  conditionName: string;
-  dimensions: Record<string, string | number> | null;
-  prioritized: number;
-  deprioritized: number;
-  neutral: number;
-  totalTrials: number;
-  selectedValueWinRate: number | null;
-  strongly: number;
-  somewhat: number;
-  opponentSomewhat: number;
-  opponentStrongly: number;
-  unknownCount: number;
-};
-
-export type DomainAnalysisVignetteDetail = {
-  definitionId: string;
-  definitionName: string;
-  definitionVersion: number;
-  aggregateRunId: string | null;
-  otherValueKey: DomainAnalysisValueKey;
-  prioritized: number;
-  deprioritized: number;
-  neutral: number;
-  totalTrials: number;
-  selectedValueWinRate: number | null;
-  conditions: DomainAnalysisConditionDetail[];
 };
 
 export type DomainAnalysisValueDetailResult = {
@@ -235,24 +211,6 @@ export type DomainTrialRunStatus = {
   modelStatuses: DomainTrialModelStatus[];
 };
 
-export type DomainAnalysisConditionTranscript = {
-  id: string;
-  runId: string;
-  scenarioId: string | null;
-  modelId: string;
-  decisionCode: string | null;
-  decisionCodeSource: string | null;
-  decisionMetadata: unknown;
-  definitionSnapshot?: unknown;
-  pairOverride?: DomainAnalysisValuePair | null;
-  decisionModelV2?: TranscriptDecisionModelResult | null;
-  turnCount: number;
-  tokenCount: number;
-  durationMs: number;
-  createdAt: Date;
-  content: unknown;
-};
-
 export type DomainAnalysisMissingReasonCode =
   | 'NO_COMPLETED_RUNS'
   | 'NO_SIGNATURE_MATCH'
@@ -274,9 +232,6 @@ export function runModelsContainAll(config: unknown, defaultModelIds: string[]):
   return defaultModelIds.every((id) => (models as unknown[]).includes(id));
 }
 
-export function isDomainAnalysisValueKey(value: string): value is DomainAnalysisValueKey {
-  return DOMAIN_ANALYSIS_VALUE_KEYS.includes(value as DomainAnalysisValueKey);
-}
 
 export function parseDomainAnalysisScoreMethod(value: string | null | undefined): DomainAnalysisScoreMethod {
   return value === 'FULL_BT' ? 'FULL_BT' : 'LOG_ODDS';
@@ -395,107 +350,6 @@ export async function resolveSignatureRuns(
   };
 }
 
-function incrementValueCount(
-  modelMap: Map<DomainAnalysisValueKey, DomainAnalysisValueCounts>,
-  valueKey: DomainAnalysisValueKey,
-  field: 'prioritized' | 'deprioritized' | 'neutral',
-): void {
-  const current = modelMap.get(valueKey) ?? { prioritized: 0, deprioritized: 0, neutral: 0 };
-  current[field] += 1;
-  modelMap.set(valueKey, current);
-}
-
-export function incrementPairwiseWin(
-  pairwiseWins: Map<DomainAnalysisValueKey, Map<DomainAnalysisValueKey, number>>,
-  winner: DomainAnalysisValueKey,
-  loser: DomainAnalysisValueKey,
-): void {
-  const winnerMap = pairwiseWins.get(winner) ?? new Map<DomainAnalysisValueKey, number>();
-  winnerMap.set(loser, (winnerMap.get(loser) ?? 0) + 1);
-  pairwiseWins.set(winner, winnerMap);
-}
-
-function getPairwiseWinCount(
-  pairwiseWins: Map<DomainAnalysisValueKey, Map<DomainAnalysisValueKey, number>>,
-  winner: DomainAnalysisValueKey,
-  loser: DomainAnalysisValueKey,
-): number {
-  return pairwiseWins.get(winner)?.get(loser) ?? 0;
-}
-
-export function computeSmoothedLogOddsScore(wins: number, losses: number): number {
-  return Math.log((wins + 1) / (losses + 1));
-}
-
-export function computeFullBTScores(
-  valueKeys: readonly DomainAnalysisValueKey[],
-  pairwiseWins: Map<DomainAnalysisValueKey, Map<DomainAnalysisValueKey, number>>,
-): Map<DomainAnalysisValueKey, number> {
-  const EPSILON = 1e-6;
-  const MAX_ITERATIONS = 500;
-  const TOLERANCE = 1e-8;
-
-  const strengths = new Map<DomainAnalysisValueKey, number>(valueKeys.map((valueKey) => [valueKey, 1]));
-
-  for (let iteration = 0; iteration < MAX_ITERATIONS; iteration += 1) {
-    const nextStrengths = new Map<DomainAnalysisValueKey, number>();
-    let maxLogDelta = 0;
-
-    for (const valueKey of valueKeys) {
-      const currentStrength = strengths.get(valueKey) ?? 1;
-      let totalWins = 0;
-      let denominator = 0;
-      let hasComparisons = false;
-
-      for (const opponent of valueKeys) {
-        if (opponent === valueKey) continue;
-        const wins = getPairwiseWinCount(pairwiseWins, valueKey, opponent);
-        const losses = getPairwiseWinCount(pairwiseWins, opponent, valueKey);
-        const matches = wins + losses;
-        if (matches <= 0) continue;
-
-        const opponentStrength = strengths.get(opponent) ?? 1;
-        const sumStrength = currentStrength + opponentStrength;
-        if (sumStrength <= 0) continue;
-
-        hasComparisons = true;
-        totalWins += wins;
-        denominator += matches / sumStrength;
-      }
-
-      let nextStrength = currentStrength;
-      if (hasComparisons && denominator > 0) {
-        nextStrength = totalWins / denominator;
-      }
-      if (!Number.isFinite(nextStrength) || nextStrength <= 0) {
-        nextStrength = EPSILON;
-      }
-      nextStrengths.set(valueKey, nextStrength);
-    }
-
-    const logValues = valueKeys.map((valueKey) => Math.log(Math.max(nextStrengths.get(valueKey) ?? EPSILON, EPSILON)));
-    const meanLog = logValues.reduce((sum, value) => sum + value, 0) / (logValues.length || 1);
-    const normalizationFactor = Math.exp(meanLog);
-
-    for (const valueKey of valueKeys) {
-      const normalized = Math.max((nextStrengths.get(valueKey) ?? EPSILON) / normalizationFactor, EPSILON);
-      const prev = Math.max(strengths.get(valueKey) ?? EPSILON, EPSILON);
-      const logDelta = Math.abs(Math.log(normalized) - Math.log(prev));
-      if (logDelta > maxLogDelta) maxLogDelta = logDelta;
-      strengths.set(valueKey, normalized);
-    }
-
-    if (maxLogDelta < TOLERANCE) break;
-  }
-
-  return new Map(
-    valueKeys.map((valueKey) => {
-      const strength = Math.max(strengths.get(valueKey) ?? EPSILON, EPSILON);
-      return [valueKey, Math.log(strength)];
-    }),
-  );
-}
-
 export async function resolveValuePairsInChunks(
   definitionIds: string[],
 ): Promise<Map<string, DomainAnalysisValuePair>> {
@@ -515,73 +369,6 @@ export async function resolveValuePairsInChunks(
   }
 
   return valuePairByDefinition;
-}
-
-
-
-export function aggregateValueCountsFromTranscripts(
-  transcripts: Array<{
-    runId: string;
-    modelId: string;
-    decisionCode: string | null;
-    decisionMetadata: unknown;
-    scenario: { orientationFlipped: boolean | null } | null;
-  }>,
-  sourceRunDefinitionById: Map<string, string>,
-  valuePairByDefinition: Map<string, DomainAnalysisValuePair>,
-): {
-  aggregatedByModel: Map<string, Map<DomainAnalysisValueKey, DomainAnalysisValueCounts>>;
-  pairwiseWinsByModel: Map<string, Map<DomainAnalysisValueKey, Map<DomainAnalysisValueKey, number>>>;
-  analyzedDefinitionIds: Set<string>;
-} {
-  const aggregatedByModel = new Map<string, Map<DomainAnalysisValueKey, DomainAnalysisValueCounts>>();
-  const pairwiseWinsByModel = new Map<string, Map<DomainAnalysisValueKey, Map<DomainAnalysisValueKey, number>>>();
-  const analyzedDefinitionIds = new Set<string>();
-
-  for (const transcript of transcripts) {
-    const definitionId = sourceRunDefinitionById.get(transcript.runId);
-    if (definitionId == null || definitionId === '') continue;
-    const pair = valuePairByDefinition.get(definitionId);
-    if (!pair) continue;
-
-    const resolved = resolveTranscriptDecisionModel({
-      decisionCode: transcript.decisionCode,
-      decisionMetadata: transcript.decisionMetadata,
-      orientationFlipped: transcript.scenario?.orientationFlipped ?? null,
-      pairOverride: pair,
-    });
-    const canonical = resolved.canonical;
-    if (canonical.direction === 'unknown') continue;
-
-    let valueMap = aggregatedByModel.get(transcript.modelId);
-    if (!valueMap) {
-      valueMap = new Map<DomainAnalysisValueKey, DomainAnalysisValueCounts>();
-      aggregatedByModel.set(transcript.modelId, valueMap);
-    }
-
-    let pairwiseWins = pairwiseWinsByModel.get(transcript.modelId);
-    if (!pairwiseWins) {
-      pairwiseWins = new Map<DomainAnalysisValueKey, Map<DomainAnalysisValueKey, number>>();
-      pairwiseWinsByModel.set(transcript.modelId, pairwiseWins);
-    }
-
-    if (
-      canonical.direction !== 'neutral'
-      && canonical.favoredValueKey != null
-      && canonical.opposedValueKey != null
-    ) {
-      incrementValueCount(valueMap, canonical.favoredValueKey, 'prioritized');
-      incrementValueCount(valueMap, canonical.opposedValueKey, 'deprioritized');
-      incrementPairwiseWin(pairwiseWins, canonical.favoredValueKey, canonical.opposedValueKey);
-    } else {
-      incrementValueCount(valueMap, pair.valueA, 'neutral');
-      incrementValueCount(valueMap, pair.valueB, 'neutral');
-    }
-
-    analyzedDefinitionIds.add(definitionId);
-  }
-
-  return { aggregatedByModel, pairwiseWinsByModel, analyzedDefinitionIds };
 }
 
 export function supportsTemperature(apiConfig: unknown): boolean {
